move Developer to Domain, before refactor
