+++ SpreadsheetUtility/SpreadsheetUtility.Library/Processors/GanttChartProcessor/GanttChartProcessor.cs
@@ -8,6 +8,7 @@ using SpreadsheetUtility.Library.Mappers;
 using SpreadsheetUtility.Library.Models;
 using SpreadsheetUtility.Library.Providers;
 using SpreadsheetUtility.Library.TaskAssigners;
+using SpreadsheetUtility.Library.Domain;
 
 namespace SpreadsheetUtility.Library.Processors;
