+++ SpreadsheetUtility/SpreadsheetUtility.Library/Processors/GanttChartProcessor/TaskAssigners/ITaskAssignmentStrategy.cs
@@ -1,5 +1,5 @@
 ﻿using SpreadsheetUtility.Library.Models;
-using SpreadsheetUtility.Library.Processors;
+using SpreadsheetUtility.Library.Domain;
 
 namespace SpreadsheetUtility.Library.TaskAssigners;
