+++ SpreadsheetUtility/SpreadsheetUtility.Test/GanttServiceTests.cs
@@ -13,6 +13,7 @@ using SpreadsheetUtility.Test.Helpers;
 using SpreadsheetUtility.Library.Grouppers;
 using SpreadsheetUtility.Library.TaskSorters;
 using Microsoft.Extensions.DependencyInjection;
+using SpreadsheetUtility.Library.Domain;
 
 namespace SpreadsheetUtility.Test
 {
